Update library to JaxLinOp name.
