+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,6 +1,6 @@
 ---
 blank_issues_enabled: false
 contact_links:
-  - name: GPJax Community Support
-    url: https://github.com/thomaspinder/GPJax/discussions
+  - name: JaxLinOp Community Support
+    url: https://github.com/JaxGaussianProcesses/JaxLinOp/discussions
     about: Please ask and answer questions here.
